refactor: move crl structs to a unified module
