+++ crypto-ffi/src/generic/context/e2ei.rs
@@ -1,10 +1,10 @@
 use std::{collections::HashMap, ops::DerefMut};
 
 use crate::{
-    CoreCryptoError, NewCrlDistributionPoints,
+    CoreCryptoError, CrlRegistration, NewCrlDistributionPoints,
     generic::{
-        Ciphersuite, ClientId, CoreCryptoResult, CrlRegistration, E2eiConversationState, E2eiDumpedPkiEnv,
-        E2eiEnrollment, MlsCredentialType, WireIdentity, context::CoreCryptoContext,
+        Ciphersuite, ClientId, CoreCryptoResult, E2eiConversationState, E2eiDumpedPkiEnv, E2eiEnrollment,
+        MlsCredentialType, WireIdentity, context::CoreCryptoContext,
     },
 };
 use core_crypto::mls::conversation::Conversation as _;
