+++ crypto-ffi/src/generic/mod.rs
@@ -28,29 +28,6 @@ use core_crypto_keystore::Connection as Database;
 pub mod context;
 mod epoch_observer;
 
-#[derive(Debug, Clone, derive_more::From, derive_more::Into)]
-pub struct NewCrlDistributionPoints(Option<Vec<String>>);
-
-uniffi::custom_newtype!(NewCrlDistributionPoints, Option<Vec<String>>);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, uniffi::Record)]
-/// Supporting struct for CRL registration result
-pub struct CrlRegistration {
-    /// Whether this CRL modifies the old CRL (i.e. has a different revocated cert list)
-    pub dirty: bool,
-    /// Optional expiration timestamp
-    pub expiration: Option<u64>,
-}
-
-impl From<core_crypto::e2e_identity::CrlRegistration> for CrlRegistration {
-    fn from(value: core_crypto::e2e_identity::CrlRegistration) -> Self {
-        Self {
-            dirty: value.dirty,
-            expiration: value.expiration,
-        }
-    }
-}
-
 #[derive(Debug, Clone, uniffi::Record)]
 pub struct ProteusAutoPrekeyBundle {
     pub id: u16,
